Add support for citations
If JavaScript is enabled, citations can be added by using the tags in
the Publications.tex file. You can for example reference a famous code
generation algorithm like this: '<a class="citation">QUILLERE00</a>'.
The relevant Bibliography will be automatically added to the end of the
page and the citation links to it. This works at least with a recent
firefox and with chromium.

If javascript is disabled, the citation is shown, but no references are
added.